refactor: modularize navbar component code and update sidebar
- refactor navbar component to modularize code
- logo related code is now a Logo component
- sidebar menu icon is now a HamburgerMenu component
- all other navbar element related code is now a NavBarElements component
- refactor sidebar related code so its related logic is simplified
- refactor sidebar CSS so it now appears from top to bottom rather than right to left
