+++ client/src/components/ui/HamburgerMenu/HamburgerMenu.css
@@ -1,0 +1,10 @@
+.navbar__menu-icon {
+    display: none;
+}
+
+@media (max-width: 768px) {
+    .navbar__menu-icon {
+        display: block;
+        cursor: pointer;
+    }
+}
