+++ client/src/components/ui/Navbar/Navbar.css
@@ -1,148 +1,75 @@
 .navbar__container {
-  /* max-width: 1200px; */
-  max-width: 90vw;
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 100%;
+    max-width: 90vw;
+    margin: 0 auto;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    height: 100%;
 }
 
 .navbar {
     padding: 5px 20px;
-    /* height: 20%; */
     background-color: #fff6e5;
     position: relative;
-    /* border: 1px solid rgba(0, 0, 0, 0.3); */
-    box-shadow: rgba(0, 0, 0, 0.2) 0 1px ;
+    box-shadow: rgba(0, 0, 0, 0.2) 0 1px;
     width: 100%;
 }
 
 .navbar__logo {
-  /* display: flex;
-  justify-content: start; */
-  max-height: 100%;
-  max-width: 100%;
+    max-height: 100%;
+    max-width: 100%;
 }
 
 .navbar__logo img {
-  max-height: 80px;
-  background-color: orange;
+    max-height: 80px;
+    background-color: orange;
 }
-
-/* .navbar__logo a{
-height: 0px;} */
-
-
-.navbar__menu-icon{
-  display:none;
-}
-
-/* .nav-elements {
-  position: absolute;
-  right: 0;
-  top: 60px;
-  background-color: #fef7e5;
-  width: 0;
-  overflow: hidden;
-  height: calc(100vh-60px);
-  transition: all 0.3s ease-in;
-} */
-
-.nav-elements ul {
-  display: flex;
-  justify-content: space-between;
-  list-style-type: none;
-}
-
-.nav-elements ul li:not(:last-child) {
-  margin-right: 30px;
-}
-
-.nav-elements ul a, .nav-elements ul button  {
-    font-family: 'Montserrat';  
-    font-size:18px;
-    font-weight: 600;
-    text-decoration: none;
-    color: #2f234f;
-    /* background: #ffa500; */
-    padding: 10px 15px;
-    /* border-radius: .25rem; */
-    /* box-shadow: 0 2px; */
-
-}
-
-.nav-elements ul a:hover {
-  font-size:18px;
-  font-weight: 600;
-  text-decoration: none;
-  color: #2f234f;
-  box-shadow: 0 2px;
-  cursor: pointer;
-}
-
-.nav-elements ul a.active {
-  color: #2f234f;
-  font-weight: 800;
-  position: relative;
-  box-shadow: 0 2px;
-}
-
-
-/* .nav-elements ul a.active::after {
-  content: "";
-  position: absolute;
-  bottom: -4px;
-  left: 0;
-  width: 100%;
-  height: 2px;
-  background-color: #575c4c;
-} */
 
 .dropdown {
     position: relative;
 }
 
-.dropdown-menu.active{
+.dropdown-menu.active {
     position: absolute;
     right: 0;
-    top: calc(100% + .25rem);
+    top: calc(100% + 0.25rem);
     background-color: white;
-    padding: .75rem;
-    border-radius: .25rem;
+    padding: 0.75rem;
+    border-radius: 0.25rem;
     box-shadow: 0 2px;
     opacity: 1;
     transform: translateY(0);
     z-index: 1;
     width: 100%;
- 
 }
 
 .dropdown-menu.active > ul,
-.dropdown-menu ul{
+.dropdown-menu ul {
     display: flex;
     flex-direction: column;
 }
 
 .dropdown-menu > ul > li,
 .dropdown-menu > ul > li > button {
-        width: 100%;
-        text-align: center;
-        padding: 10px 5px;
+    width: 100%;
+    text-align: center;
+    padding: 10px 5px;
 }
 
-.dropdown-menu{
+.dropdown-menu {
     width: 100%;
     position: absolute;
     right: 0;
     top: calc(100% + 0.25rem);
     background-color: white;
-    padding: .75rem;
-    border-radius: .25rem;
+    padding: 0.75rem;
+    border-radius: 0.25rem;
     box-shadow: 0 2px;
     opacity: 0;
     transform: translateY(-5px);
-    transition: opacity 150ms ease-in-out, transform 150ms ease-in-out;
+    transition:
+        opacity 150ms ease-in-out,
+        transform 150ms ease-in-out;
 }
 
 /* .profile__dropdown.false {
@@ -174,51 +101,8 @@ height: 0px;} */
     transform: rotate(45deg);
 } */
 
-
-
-@media (max-width: 768px) {
-  .navbar-elements ul li:not(:last-child){
-    margin-right: 30px;
-  }
-}
-
 @media (max-width: 768px) {
     .navbar {
         width: 100%;
     }
-    .navbar__menu-icon {
-        display: block;
-        cursor: pointer;
-    }
-
-    .nav-elements {
-        display: flex;
-        position: absolute;
-        right: 0;
-        top: 60px;
-        width: 0px;
-        background-color: #fef7e5;
-        height: calc(100vh-60px);
-        transition: all 0.3s ease-in;
-        overflow:hidden;
-        z-index: 1;
-        border-radius: 5px;
-        align-items: center;
-        justify-content: center;
-        padding-bottom: 30px;
-    }
-
-    .nav-elements.active{
-        width: 270px;
-    }
-
-    .nav-elements ul{
-        display: flex;
-        flex-direction: column;
-    }
-
-    .nav-elements ul li {
-        margin-right: unset;
-        margin-top: 22px;
-    }
-    }
+}
